refactor: update collector service interfaces (#1253)
* refactor: update collector service interfaces


* refactor: add collector api parameter types and rename to model.ts


---------
