+++ apps/web/src/services/asset-inventory/collector/collector-main/CollectorPage.vue
@@ -66,7 +66,7 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 import { useCollectorPageStore } from '@/services/asset-inventory/collector/collector-main/collector-page-store';
 import CollectorContents from '@/services/asset-inventory/collector/collector-main/modules/CollectorContents.vue';
 import CollectorNoData from '@/services/asset-inventory/collector/collector-main/modules/CollectorNoData.vue';
-import { COLLECTOR_QUERY_HELPER_SET } from '@/services/asset-inventory/collector/type';
+import { COLLECTOR_QUERY_HELPER_SET } from '@/services/asset-inventory/collector/model';
 import ProviderList from '@/services/asset-inventory/components/ProviderList.vue';
 import { ASSET_INVENTORY_ROUTE } from '@/services/asset-inventory/route-config';
